More Refactoring Changes
- delete jdt.ls LinkedProposalModel and replace all usage
  with jdt.core.manipulation LinkedProposalModelCore
- delete jdt.ls LinkedProposalPositionGroup and replace all usage
  with jdt.core.manipulation LinkedProposalPositionGroupCore
