+++ org.eclipse.jdt.ls.core/src/org/eclipse/jdt/ls/core/internal/corext/refactoring/code/ExtractMethodRefactoring.java
@@ -96,6 +96,8 @@ import org.eclipse.jdt.internal.core.refactoring.descriptors.RefactoringSignatur
 import org.eclipse.jdt.internal.corext.dom.ASTNodes;
 import org.eclipse.jdt.internal.corext.dom.Bindings;
 import org.eclipse.jdt.internal.corext.dom.Selection;
+import org.eclipse.jdt.internal.corext.fix.LinkedProposalModelCore;
+import org.eclipse.jdt.internal.corext.fix.LinkedProposalPositionGroupCore;
 import org.eclipse.jdt.internal.corext.util.JdtFlags;
 import org.eclipse.jdt.ls.core.internal.BindingLabelProvider;
 import org.eclipse.jdt.ls.core.internal.Messages;
@@ -105,8 +107,6 @@ import org.eclipse.jdt.ls.core.internal.corext.dom.ASTNodeFactory;
 import org.eclipse.jdt.ls.core.internal.corext.dom.BodyDeclarationRewrite;
 import org.eclipse.jdt.ls.core.internal.corext.dom.LinkedNodeFinder;
 import org.eclipse.jdt.ls.core.internal.corext.dom.StatementRewrite;
-import org.eclipse.jdt.ls.core.internal.corext.fix.LinkedProposalModel;
-import org.eclipse.jdt.ls.core.internal.corext.fix.LinkedProposalPositionGroup;
 import org.eclipse.jdt.ls.core.internal.corext.refactoring.Checks;
 import org.eclipse.jdt.ls.core.internal.corext.refactoring.JDTRefactoringDescriptorComment;
 import org.eclipse.jdt.ls.core.internal.corext.refactoring.JavaRefactoringArguments;
@@ -162,7 +162,7 @@ public class ExtractMethodRefactoring extends Refactoring {
 	private ASTNode fDestination;
 	// either of type TypeDeclaration or AnonymousClassDeclaration
 	private ASTNode[] fDestinations;
-	private LinkedProposalModel fLinkedProposalModel;
+	private LinkedProposalModelCore fLinkedProposalModel;
 
 	private static final String EMPTY = ""; //$NON-NLS-1$
 
@@ -292,7 +292,7 @@ public class ExtractMethodRefactoring extends Refactoring {
 		fRoot = astRoot;
 	}
 
-	public void setLinkedProposalModel(LinkedProposalModel linkedProposalModel) {
+	public void setLinkedProposalModel(LinkedProposalModelCore linkedProposalModel) {
 		fLinkedProposalModel = linkedProposalModel;
 	}
 
@@ -535,7 +535,7 @@ public class ExtractMethodRefactoring extends Refactoring {
 			MethodDeclaration mm = createNewMethod(selectedNodes, fCUnit.findRecommendedLineSeparator(), substituteDesc);
 
 			if (fLinkedProposalModel != null) {
-				LinkedProposalPositionGroup typeGroup = fLinkedProposalModel.getPositionGroup(KEY_TYPE, true);
+				LinkedProposalPositionGroupCore typeGroup = fLinkedProposalModel.getPositionGroup(KEY_TYPE, true);
 				typeGroup.addPosition(fRewriter.track(mm.getReturnType2()), false);
 
 				ITypeBinding typeBinding = fAnalyzer.getReturnTypeBinding();
@@ -546,7 +546,7 @@ public class ExtractMethodRefactoring extends Refactoring {
 					}
 				}
 
-				LinkedProposalPositionGroup nameGroup = fLinkedProposalModel.getPositionGroup(KEY_NAME, true);
+				LinkedProposalPositionGroupCore nameGroup = fLinkedProposalModel.getPositionGroup(KEY_NAME, true);
 				nameGroup.addPosition(fRewriter.track(mm.getName()), false);
 
 				ModifierCorrectionSubProcessor.installLinkedVisibilityProposals(fLinkedProposalModel, fRewriter, mm.modifiers(), false);
@@ -999,7 +999,7 @@ public class ExtractMethodRefactoring extends Refactoring {
 			arguments.add(ASTNodeFactory.newName(fAST, getMappedName(duplicate, parameter)));
 		}
 		if (fLinkedProposalModel != null) {
-			LinkedProposalPositionGroup nameGroup = fLinkedProposalModel.getPositionGroup(KEY_NAME, true);
+			LinkedProposalPositionGroupCore nameGroup = fLinkedProposalModel.getPositionGroup(KEY_NAME, true);
 			nameGroup.addPosition(fRewriter.track(invocation.getName()), false);
 		}
